Revert "Add some ladderSystem match function"
