+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
 project(src)
 include_directories(../lib)
 
-add_library(ACGMain ACGMain.cpp Printer.cpp Reader.cpp AccountSystem.cpp Router.cpp LadderSystem.cpp)
+add_library(ACGMain ACGMain.cpp Printer.cpp Reader.cpp AccountSystem.cpp Router.cpp )
