+++ src/LadderSystem.h
@@ -9,25 +9,18 @@
 #include <string>
 #include "AccountSystem.h"
 
-enum LadderLevel{
-    COPPER = 0, SILVER = 1, GOLD = 2, DIAMOND = 3, GREAT_MASTER = 4
+enum LEVEL{
+    COPPER = 0, SILVER = 1, GOLD = 2, DIAMOND = 3, GREAT_MASTER = ,
 };
-class LadderSystem;
-class LadderCalculate {
-    public:
-    static LadderLevel get_level(int point);
-};
-
 class LadderPlayer {
-public:
+    LadderPlayer(int u_id, std::string id, int point_) : u_id_(u_id), id_(id), point_(point){
+        level = LadderSystem::get_level(point_);
+    }
+    int u_id_;
     std::string id_;
     int point_;
     int Last_game;
-    LadderLevel level;
-    int u_id_;
-    LadderPlayer(int u_id, std::string name, int point) : u_id_(u_id), id_(name), point_(point) {
-        level = LadderCalculate::get_level(point);
-    }
+    LEVEL level;
 };
 
 class LadderSystem {
@@ -35,10 +28,34 @@ public:
     std::vector<LadderPlayer *> &participant;
     void join(Account *account);
     void match();
-    void leave(int remove_id);
+    LEVEL get_level(int point);
 private:
 
 };
+void LadderSystem::join(Account *account) {
+    int ladder_point = account->get_ladder_point();
+    std:string id = account->get_name();
+    int u_num = account->get_unique_num();
+    participant.push_back(new LadderPlayer(u_num, id, ladder_point));
+}
+LEVEL LadderSystem::get_level(int point) {
+    if (point < 2000) {
+        return COPPER; // 0 - 1999
+    } else if (point < 2500) {
+        return SILVER; // 2000 - 2499
+    } else if (point < 3000) {
+        return GOLD;   // 2500 - 2999
+    } else if (point < 3500) {
+        return DIAMOND; // 3000 - 3499
+    } else {
+        return GREAT_MASTER;
+    }
+}
 
+
+void LadderSystem::match() {
+    // TODO
+
+}
 
 #endif //ANOTHER_CARD_GAME_ACG_LADDERSYSTEM_H
